python-backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
langchain==0.1.9
langchain-openai==0.0.7
langchain-pinecone==0.0.3
pinecone-client==3.0.2
supabase>=0.7.1
langchain-community>=0.0.1